move fft to their own module
